+++ apps/typegpu-docs/src/components/FpsCounter.tsx
